rename files. simplify inheritence
