+++ src/element-observer.js
@@ -1,35 +1,19 @@
-import { Observer, ObserverInterface } from './observer-interface'
+import Observer from './observer'
 
-function isElementInViewport(element, offset, viewportState) {
-  const rect = element.getBoundingClientRect()
-  return (
-    !!(rect.width && rect.height) &&
-    rect.top < viewportState.height + offset &&
-    rect.bottom > 0 - offset &&
-    rect.left < viewportState.width + offset &&
-    rect.right > 0 - offset
-  )
-}
-
-function isElementInDOM(element) {
-  return element && element.parentNode
-}
-
-const ElementObserver = ObserverInterface(function ElementObserver(element, opts = {}) {
-  if (!(this instanceof ElementObserver)) {
-    return new ElementObserver(...arguments)
-  }
+export default function ElementObserver(element, opts = {}) {
+  if (!(this instanceof ElementObserver)) return new ElementObserver(...arguments)
 
   this.element = element
   this.onEnter = opts.onEnter
   this.onExit = opts.onExit
   this._didEnter = false
-  const viewport = Observer.call(this, opts)
 
-  if (isElementInDOM(element)) {
-    this.check(viewport.getState())
-  }
-})
+  const viewport = Observer.call(this, opts)
+  if (isElementInDOM(element)) this.check(viewport.getState())
+}
+
+ElementObserver.prototype = Object.create(Observer.prototype)
+ElementObserver.prototype.constructor = ElementObserver
 
 ElementObserver.prototype.check = function(viewportState) {
   const { onEnter, onExit, element, offset, once, _didEnter } = this
@@ -53,4 +37,17 @@ ElementObserver.prototype.check = function(viewportState) {
   }
 }
 
-export default ElementObserver
+function isElementInViewport(element, offset, viewportState) {
+  const rect = element.getBoundingClientRect()
+  return (
+    !!(rect.width && rect.height) &&
+    rect.top < viewportState.height + offset &&
+    rect.bottom > 0 - offset &&
+    rect.left < viewportState.width + offset &&
+    rect.right > 0 - offset
+  )
+}
+
+function isElementInDOM(element) {
+  return element && element.parentNode
+}
