+++ src/observer.js
@@ -1,10 +1,10 @@
 import { addViewportObserver, removeViewportObserver } from './viewports-manager'
 
 /**
- * @interface Observer
+ * @class Observer
  * Each type of observer implements these options/methods
  */
-export function Observer(opts) {
+export default function Observer(opts) {
   this.offset = ~~opts.offset || 0
   this.container = opts.container || document.body
   this.once = Boolean(opts.once)
@@ -18,10 +18,4 @@ Observer.prototype = {
   destroy() {
     removeViewportObserver(this)
   }
-}
-
-export function ObserverInterface(Subclass) {
-  Subclass.prototype = Object.create(Observer.prototype)
-  Subclass.prototype.constructor = Subclass
-  return Subclass
 }
